server/requirements.txt
pyserial-asyncio
Pillow
webuntis
pygame
